refactor(courses): separate the logic
